+++ crates/exchange-rates/src/feeds.rs
