+++ crates/exchange-rates/src/lib.rs
@@ -1,7 +1,7 @@
 mod init;
 mod types;
 mod utils;
-pub mod feed;
+pub mod feeds;
 pub mod error;
 pub mod commands;
